convert change configuration as well
