--- src/lib/ChargeStation/eventHandlers/ocpp-16/handle-change-configuration.js
@@ -1,15 +1,0 @@
-export default async function handleChangeConfiguration({
-  chargepoint,
-  callMessageId,
-  callMessageBody,
-}) {
-  const { key } = callMessageBody;
-
-  chargepoint.configuration.setVariable(key, callMessageBody);
-
-  const response = {
-    status: 'Accepted',
-  };
-
-  chargepoint.writeCallResult(callMessageId, response);
-}
