+++ src/lib/ChargeStation/eventHandlers/ocpp-16/handle-change-configuration.ts
@@ -1,0 +1,17 @@
+import { ChargeStationEventHandler } from 'lib/ChargeStation/eventHandlers';
+import { ChangeConfigurationRequest } from 'schemas/ocpp/1.6/ChangeConfiguration';
+
+const handleChangeConfigurationHandler: ChargeStationEventHandler<ChangeConfigurationRequest> =
+  ({ chargepoint, callMessageId, callMessageBody }) => {
+    const { key } = callMessageBody;
+
+    chargepoint.configuration.setVariable(key, callMessageBody);
+
+    const response = {
+      status: 'Accepted',
+    };
+
+    chargepoint.writeCallResult(callMessageId, response);
+  };
+
+export default handleChangeConfigurationHandler;
